Java Script Code
The Java Script 
I'm going to begin by commenting out the User constructor. To create a class, I need to begin with the class keyword. Class User, note that the User is capitalized, just like the constructor function above. Now, inside of the class, I need to invoke the constructor. The constructor will take two parameters, userName and email. I will initialize the class properties, this.userName=userName, and this.email=email. The class so far looks very similar in structure to the object constructor.